Add tooltip to Edit chart
